Add registration and roles privileges

+++ src/main/java/com/optimal/solution/controller/CommentController.java
@@ -1,6 +1,7 @@
 package com.optimal.solution.controller;
 
 import com.optimal.solution.dto.CommentDto;
+import com.optimal.solution.model.Comment;
 import com.optimal.solution.service.CommentService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
@@ -8,6 +9,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/comments")
@@ -19,7 +22,7 @@ public class CommentController {
     @GetMapping(value = {"", "/"}, produces = "application/json")
     ResponseEntity<?> findAll() {
         try {
-            LOGGER.info("Getting list of comments!");
+            LOGGER.info("Getting list of comments");
             return ResponseEntity.ok(commentService.findAll());
         } catch (Exception e) {
             LOGGER.error("Error with getting list of comments!", e);
@@ -30,8 +33,13 @@ public class CommentController {
     @GetMapping(value = {"/{id}"}, produces = "application/json")
     ResponseEntity<?> findById(@PathVariable int id) {
         try {
-            LOGGER.info("Getting comment by id");
-            return ResponseEntity.ok(commentService.findById(id));
+            Optional<Comment> comment = commentService.findById(id);
+            if (comment.isPresent()) {
+                LOGGER.info("Getting comment by id");
+                return ResponseEntity.ok(comment);
+            } else {
+                return ResponseEntity.ok("No comment found with id " + id);
+            }
         } catch (Exception e) {
             LOGGER.error("Error with getting comment by id!", e);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -52,8 +60,12 @@ public class CommentController {
     @DeleteMapping(value = {"/{id}"}, produces = "application/json")
     ResponseEntity<?> deleteById(@PathVariable int id) {
         try {
-            LOGGER.info("Deleting comment by id");
-            return ResponseEntity.ok(commentService.deleteById(id));
+            if (commentService.findById(id).isPresent()) {
+                LOGGER.info("Deleting comment by id");
+                return ResponseEntity.ok(commentService.deleteById(id));
+            } else {
+                return ResponseEntity.ok("No comment found with id " + id);
+            }
         } catch (Exception e) {
             LOGGER.error("Error with deleting comment by id!", e);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
