+++ src/main/java/com/optimal/solution/controller/PostController.java
@@ -1,6 +1,7 @@
 package com.optimal.solution.controller;
 
 import com.optimal.solution.dto.PostDto;
+import com.optimal.solution.model.Post;
 import com.optimal.solution.service.PostService;
 import lombok.RequiredArgsConstructor;
 import org.slf4j.Logger;
@@ -8,6 +9,8 @@ import org.slf4j.LoggerFactory;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/posts")
@@ -31,7 +34,12 @@ public class PostController {
     ResponseEntity<?> findById(@PathVariable int id) {
         try {
             LOGGER.info("Getting post by id");
-            return ResponseEntity.ok(postService.findById(id));
+            Optional<Post> post = postService.findById(id);
+            if(post.isPresent()){
+                return ResponseEntity.ok(post);
+            } else {
+                return ResponseEntity.ok("No post found with id " + id);
+            }
         } catch (Exception e) {
             LOGGER.error("Error with getting post by id!", e);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
@@ -52,8 +60,12 @@ public class PostController {
     @DeleteMapping(value = {"/{id}"}, produces = "application/json")
     ResponseEntity<?> deleteById(@PathVariable int id) {
         try {
-            LOGGER.info("Deleting category by id");
-            return ResponseEntity.ok(postService.deleteById(id));
+            LOGGER.info("Deleting post by id");
+            if(postService.findById(id).isPresent()){
+                return ResponseEntity.ok(postService.deleteById(id));
+            } else {
+                return ResponseEntity.ok("No post found with id " + id);
+            }
         } catch (Exception e) {
             LOGGER.error("Error with deleting post by id!", e);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
